requirements.txt
boto3==1.4.7
botocore==1.7.12
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-s3-storage==0.12.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
docutils==0.14
futures==3.1.1
idna==2.6
jmespath==0.9.3
psycopg2==2.7.3.1
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
s3transfer==0.1.11
six==1.10.0
urllib3==1.22
xxhash==1.0.1
